// src/app/api/model/problemDetail.ts
/**
 * Goaldone API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { FieldError } from './fieldError';


/**
 * Fehlerantwort nach RFC 9457 (Problem Details for HTTP APIs)
 */
export interface ProblemDetail { 
    /**
     * URI, die den Fehlertyp eindeutig identifiziert
     */
    type: string;
    /**
     * Kurze, menschenlesbare Fehlerbeschreibung
     */
    title: string;
    /**
     * HTTP Status Code
     */
    status: number;
    /**
     * Detaillierte Fehlermeldung für diesen spezifischen Aufruf
     */
    detail: string;
    /**
     * URI des Endpunkts, der den Fehler verursacht hat
     */
    instance: string;
    /**
     * Feldspezifische Validierungsfehler (nur bei 400 Bad Request)
     */
    errors?: Array<FieldError>;
}